+++ studio/CHANGELOG.md
@@ -4,6 +4,14 @@ Binaries are attached to the github release otherwise all images can be found [h
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.180.5](https://github.com/wundergraph/cosmo/compare/studio@0.180.4...studio@0.180.5) (2026-09-09)
+
+**Note:** Version bump only for package studio
+
+## [0.180.4](https://github.com/wundergraph/cosmo/compare/studio@0.180.3...studio@0.180.4) (2026-09-09)
+
+**Note:** Version bump only for package studio
+
 ## [0.180.3](https://github.com/wundergraph/cosmo/compare/studio@0.180.2...studio@0.180.3) (2026-09-08)
 
 **Note:** Version bump only for package studio